Correcoes alterar timesheet

+++ src/model/dao/TimesheetDAO.js
@@ -44,28 +44,33 @@ class TimesheetDAO {
     }
 
     async alterarTimesheet(timesheet) {
-        
-        await Apontamento.destroy({
-            where: { id_timesheet: timesheet.id }
-        });
-        
-        await Timesheet.update({
-            data: timesheet.data,
-            id_usuario: timesheet.usuario.id,
-            id_cliente: timesheet.cliente.id,
-            observacao: timesheet.observacao,
-        }, {
-            where: { id: timesheet.id }
-        });
-        
-        timesheet.apontamentos.map(async a => {
-            await Apontamento.create({
-                sequencia: a.sequencia,
-                hora: a.hora,
-                id_timesheet: timesheet.id
-            });
-        });
 
+        if(timesheet.cliente && timesheet.usuario){
+        
+            await Apontamento.destroy({
+                where: { id_timesheet: timesheet.id }
+            });
+            
+            await Timesheet.update({
+                data: timesheet.data,
+                id_usuario: timesheet.usuario.id,
+                id_cliente: timesheet.cliente.id,
+                observacao: timesheet.observacao,
+            }, {
+                where: { id: timesheet.id }
+            });
+
+            if(timesheet.apontamentos.length > 0){
+            
+                timesheet.apontamentos.map(async a => {
+                    await Apontamento.create({
+                        sequencia: a.sequencia,
+                        hora: a.hora,
+                        id_timesheet: timesheet.id
+                    });
+                });
+            }
+        }
     }
 
     async excluirTimesheet(timesheet) {
